fix pointer stuff

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: flfische <flfische@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/04 17:47:11 by flfische          #+#    #+#             */
-/*   Updated: 2024/03/09 13:28:25 by flfische         ###   ########.fr       */
+/*   Updated: 2024/03/18 11:03:20 by flfische         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -84,6 +84,6 @@ t_list				*ft_lstmap(t_list *lst, void *(*f)(void *),
 						void (*del)(void *));
 
 // EXTRA
-int					ft_putnbr_base(unsigned long long nbr, char *base);
+int					ft_putnbr_base(long nbr, char *base);
 
 #endif
